requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6
opencv-python==4.8.1.78
opencv-contrib-python==4.11.0.86
mediapipe==0.10.8
moviepy==1.0.3
google-generativeai==0.3.2
elevenlabs==0.2.27
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
aiofiles==23.2.1 